+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Added
+
+- Check that equality/inequality matrices/vectors are provided consistently
+- Lower and upper bound keyword arguments ``lb`` and ``ub``
+
+### Fixed
+
+- Relaxed offset check in [test\_solvers.py](examples/test_solvers.py)
+
 ## [1.2.1] - 2020/05/16
 
 ### Added